[TASK] Move up `getAllRuleSets` to `CSSBlockList`
Move this method up without any changes to the method or its callers.

We'll clean this up in later changes.

Part of #994.
